feat: expand and collapse layout panel (#915)
Add a button that toggles the layout between expanded and collapsed state. Both expanded and collapsed will grow and shrink according to content in the Layout, but in collapsed state, the max it can grow to is 20vh. In expanded state, it will grow to fill 100% of the parent container if needed. Once that max has been reached, then a scrollbar is shown. 

Implements TECH-957