fix: avoid storing incomplete start-end dates
Since ui in store is updated as soon as a date is selected in either
start or end date fields, we don't want to store incomplete start-end
dates.
I'm not sure this is the best solution, but we need some sort of
internal state that is initialised with the value passed from outside
and sets the value back only when both start and end date are defined.