fix: prevent extra api request and render of vis table when sorting and updating (#40)
Implements TECH-1033

sortField, sortDirection and page number need to be stored in redux store since multiple components change the values (sorting buttons on table, and the Update button). With this change, the sorting and current data in store are coordinated so that the api request in useAnalyticsData only triggers once.