fix: move pagination/sorting state to plugin TECH-1177
This causes the sorting information to not be stored at app level, thus
making it not possible to use the same sorting when downloading.
With Jan we concluded that since the sorting cannot be saved at the
moment, it's ok if the download does not reflect what the user has in
the main visualization.

This solves 2 issues:

1) because the plugin does not use the app's Redux store for
pagination/sorting, when paginating/sorting the table in the Interpretations modal nothing happens in the main visualization table.

2) because the Visualization component will eventually be the plugin, it
cannot rely on anything in the app, otherwise some things might not work
when the plugin is used in a dashboard.