fix: untie Plugin component from the app's Redux store (#168)
The plugin cannot rely on the app's Redux store.
Actions that work directly on the app's store need to be replaced with callbacks.
A callback can also act as feature toggle, we can choose to not pass the callback and the correspondent feature is then disabled. For example the column header click that opens the dimension modal should be disabled in dashboard app.