fix: when in the PeriodDimension, use the raw dimensionIds (#487)
Changes in this PR:

* Fix the bug where the dimension items repeatedly get prefixed with the programId,
such as: programId.programId.programId.dimensionId
* Fix the bug where tooltip content was showing the metadata id rather than the
name of the period. For the tooltip content, the metadata should just be THIS_MONTH,not
programId.THIS_MONTH, since the periods will have the same name regardless of program.
* The analytics query should add all the date conditions to a single string in the parameters
list (see unit test). The only actual code change for this was the addition of this line:
https://github.com/dhis2/line-listing-app/pull/487/files#diff-6ab44e5fc58d04348d840a9b86b9a5aa98750e705ae362de559fc5c926f21c5cR41
* Add time dimension checks to existing cypress test.