test: move tooltip open in tooltip assertion helper function
There were issues with tooltip Cypress tests after upgrading @dhis2/ui.
Most likely because now the Tooltip stays open until a mouseout event
occours and this can cover an element which is targeted for another
test.
The Esc key can also be used to close it, and this is what this change
is doing.