fix: add hash routing and plugin support for the upcoming global shell [DHIS2-19061] (#660)
To enable shareable app deep links and to maintain hash route location, the Global Shell
watches for popstate events inside client apps. To make listenable events for the global shell,
the main history.listen() listener now publishes a popstate event that can be watched.

Also adds a fix for loading the app as a plugin in Dashboard within the global shell