fix: fetch original visualization before renaming so as not to overwrite newer changes (#681) [DHIS2-19506]
Fixes DHIS2-19506

Description
Rename: fetch the visualization including subscribers, and apply name and
description to that object, which is then used in the rename PUT request.
In the app, only name and description are updated, which means the app's copy
of the visualization could still be out of date. But that's a long-standing problem
that will have to be addressed later.

Save: fetch just the visualization subscribers and apply that, to the changed visualization.